requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.8
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.0
jinja2>=3.1.0
